+++ packages/date-picker/src/interface.tsx
@@ -25,8 +25,10 @@ export type DisabledTimeProps = {
 export type RangePickerMode = "date" | "month" | "week" | "year" | "quarter"
 
 export interface PickerProps
-  extends Omit<HTMLAttributes<HTMLDivElement>,
-    "defaultValue" | "prefix" | "placeholder" | "onChange" | "onSelect">,
+  extends Omit<
+      HTMLAttributes<HTMLDivElement>,
+      "defaultValue" | "prefix" | "placeholder" | "onChange" | "onSelect"
+    >,
     BoxProps {
   allowClear?: boolean
   position?: PickerPosition
@@ -68,32 +70,39 @@ export interface DatePickerProps extends PickerProps, CommonSingleProps {
   disabledTime?: (current?: Dayjs) => DisabledTimeProps
 }
 
-export interface MonthPickerProps extends PickerProps, CommonSingleProps {
-}
+export interface MonthPickerProps extends PickerProps, CommonSingleProps {}
 
-export interface YearPickerProps extends PickerProps, CommonSingleProps {
-}
+export interface YearPickerProps extends PickerProps, CommonSingleProps {}
 
-export interface CommonRangeProps
-  extends Omit<PickerProps,
-    "onChange" | "defaultValue" | "onOk" | "defaultPickerValue" | "onSelect"> {
+export interface DateRangePickerProps extends Omit<
+  PickerProps,
+  "onChange" | "defaultValue" | "onOk" | "defaultPickerValue" | "onSelect"
+  > {
+  order?: boolean
+  disableConfirm?: boolean
   disabled?: boolean | boolean[]
   format?: string | ((value: Dayjs) => string)
-  onChange?: (dateString: string[], date: Dayjs[]) => void
-  onSelect?: (dateString: string[], date: Dayjs[]) => void
   defaultValue?: DatePickerCalendarValue[]
   value?: DatePickerCalendarValue[]
   mode?: RangePickerMode
   showTime?: boolean | RangePickerProps
   placeholder?: string[]
   timepickerProps?: RangePickerProps
-  onOk?: (dateString: string[], date: Dayjs[]) => void
-  disabledTime?: (current: Dayjs, type: "start" | "end") => DisabledTimeProps
   defaultPickerValue?: DatePickerCalendarValue[]
+  disabledTime?: (current: Dayjs, type: "start" | "end") => DisabledTimeProps
+  onOk?: (dateString: string[], date: Dayjs[]) => void
+  onChange?: (dateString: string[], date: Dayjs[]) => void
+  onSelect?: (dateString: string[], date: Dayjs[]) => void
+}
+
+export interface CommonRangeProps
+  extends DateRangePickerProps {
+
 }
 
 export interface CommonProps
-  extends Omit<PickerProps,
+  extends Omit<
+    PickerProps,
     | "prefix"
     | "placeholder"
     | "shortcuts"
@@ -108,8 +117,8 @@ export interface CommonProps
     | "onSelect"
     | "onChange"
     | "onOk"
-    | "onClear"> {
-}
+    | "onClear"
+  > {}
 
 export interface RenderSinglePickerProps extends Partial<DatePickerProps> {
   type: "day" | "month" | "year"
