+++ packages/date-picker/stories/date-picker.stories.tsx
@@ -7,7 +7,7 @@ import {
   YearPicker,
   YearPickerProps,
   DateRangePicker,
-  CommonRangeProps,
+  DateRangePickerProps,
 } from "../src"
 
 export default {
@@ -49,14 +49,9 @@ const YearTemplate: Story<YearPickerProps> = (args) => {
   return <YearPicker w={"280px"} {...args} />
 }
 
-const RangeTemplate: Story<CommonRangeProps> = (args) => {
-  return <DateRangePicker w={"320px"} {...args} />
-}
 
 export const date = DateTemplate.bind({})
 
 export const Month = MonthTemplate.bind({})
 
 export const Year = YearTemplate.bind({})
-
-export const Range = RangeTemplate.bind({})
